chore: Move `send_message_to_agent` (#665)
